+++ components/Navbar.jsx
@@ -4,21 +4,21 @@ function Navbar() {
     <nav className="w-full font-nav mb-[40px]">
       <ul className="flex basis-auto h-[48px] w-full p-0 m-0">
         <ActiveLink activeClassName="active" href="/">
-          <li className="flex justify-center items-center grow shrink basis-auto transition-all duration-1000 ease-out border-solid border border-[#111827] rounded-[64px] cursor-pointer">
+          <li className="flex justify-center items-center grow shrink basis-auto transition-all duration-400 ease-out border-solid border border-[#111827] rounded-[64px] cursor-pointer">
             <a className="flex">
               OM OSS <div className="navDot"></div>
             </a>
           </li>
         </ActiveLink>
         <ActiveLink activeClassName="active" href="nyheter">
-          <li className="flex justify-center items-center grow shrink basis-auto transition-all duration-1000 ease-out border-solid border border-[#111827] rounded-[64px] cursor-pointer">
+          <li className="flex justify-center items-center grow shrink basis-auto transition-all duration-400 ease-out border-solid border border-[#111827] rounded-[64px] cursor-pointer">
             <a className="flex">
               NYHETER <div className="navDot"></div>
             </a>
           </li>
         </ActiveLink>
         <ActiveLink activeClassName="active" href="/kontakt">
-          <li className="flex justify-center items-center grow shrink basis-auto transition-all duration-1000 ease-out border-solid border border-[#111827] rounded-[64px] cursor-pointer">
+          <li className="flex justify-center items-center grow shrink basis-auto transition-all duration-400 ease-out border-solid border border-[#111827] rounded-[64px] cursor-pointer">
             <a className="flex">
               KONTAKT <div className="navDot"></div>
             </a>
