--- pages/about.jsx
@@ -1,10 +1,0 @@
-import AboutOffer from "../components/about/AboutOffer";
-import Layout from "../components/Layout";
-
-export default function About() {
-  return (
-    <Layout>
-      <AboutOffer />
-    </Layout>
-  );
-}
