+++ pages/kontakt.js
@@ -24,22 +24,24 @@ export default function contactPage({ contactDetails }) {
           description: "Kontaktuppgifter",
         }}
       >
-        <Contact
-          phoneNumber={contactDetails[0].fields.phoneNumber}
-          email={contactDetails[0].fields.email}
-          city={contactDetails[0].fields.city}
-          zipCode={contactDetails[0].fields.zipCode}
-          streetAddress={contactDetails[0].fields.streetAddress}
-          googleMaps={contactDetails[0].fields.googleMaps}
-          instagram={contactDetails[0].fields.instagram}
-          openingHours1={contactDetails[0].fields.openingHours1}
-          openingHours2={contactDetails[0].fields.openingHours2}
-          openingHours3={contactDetails[0].fields.openingHours3}
-          openingHoursTitle1={contactDetails[0].fields.openingHoursTitle1}
-          openingHoursTitle2={contactDetails[0].fields.openingHoursTitle2}
-          openingHoursTitle3={contactDetails[0].fields.openingHoursTitle3}
-          googleMapsEmbedded={contactDetails[0].fields.googleMapsEmbedded}
-        />
+        <main>
+          <Contact
+            phoneNumber={contactDetails[0].fields.phoneNumber}
+            email={contactDetails[0].fields.email}
+            city={contactDetails[0].fields.city}
+            zipCode={contactDetails[0].fields.zipCode}
+            streetAddress={contactDetails[0].fields.streetAddress}
+            googleMaps={contactDetails[0].fields.googleMaps}
+            instagram={contactDetails[0].fields.instagram}
+            openingHours1={contactDetails[0].fields.openingHours1}
+            openingHours2={contactDetails[0].fields.openingHours2}
+            openingHours3={contactDetails[0].fields.openingHours3}
+            openingHoursTitle1={contactDetails[0].fields.openingHoursTitle1}
+            openingHoursTitle2={contactDetails[0].fields.openingHoursTitle2}
+            openingHoursTitle3={contactDetails[0].fields.openingHoursTitle3}
+            googleMapsEmbedded={contactDetails[0].fields.googleMapsEmbedded}
+          />
+        </main>
       </Layout>
     </>
   );
