+++ styles/globals.css
@@ -4,12 +4,12 @@
 @font-face {
   font-family: "IBM Plex Mono";
   font-weight: 400;
-  src: url("../public/fonts/IBMPlexMono-Regular.woff2") format("woff");
+  src: url("../public/fonts/IBMPlexMono-Regular.woff2") format("woff2");
 }
 @font-face {
-  font-family: "IBM Plex Mono";
+  font-family: "hej";
   font-weight: 700;
-  src: url("../public/fonts/IBMPlexMono-Bold.woff2") format("woff");
+  src: url("../public/fonts/IBMPlexMono-Bold.woff2") format("woff2");
 }
 @font-face {
   font-family: "Space Mono";
@@ -23,13 +23,24 @@ html,
 body {
   padding: 0;
   margin: 0;
+  font-family: "IBM Plex Mono";
 }
-#__next {
+main {
   max-width: 1920px;
-  margin: auto;
+  margin-left: 7%;
+  margin-right: 7%;
+}
+@media (min-width: 1920px) {
+  main {
+    margin: auto;
+  }
 }
 h1 {
   font-size: 28px;
+}
+
+h2 {
+  font-family: "hej";
 }
 a {
   color: inherit;
